refactor: removed api folder in handlers
