+++ internal/servers/http/handlers/auth/register.go
